reorganize serializer
