+++ SpiffWorkflow/bpmn/serializer/helpers/dictionary.py
