+++ SpiffWorkflow/bpmn/serializer/helpers/task_spec.py
@@ -1,13 +1,10 @@
 from functools import partial
 
-from uuid import UUID
-from datetime import datetime, timedelta
-
 from SpiffWorkflow.bpmn.specs.BpmnProcessSpec import BpmnDataSpecification
 
 from .dictionary import DictionaryConverter
 
-from ..specs.events.event_definitions import (
+from ...specs.events.event_definitions import (
     NoneEventDefinition,
     MultipleEventDefinition, 
     SignalEventDefinition, 
@@ -23,39 +20,8 @@ from ..specs.events.event_definitions import (
     NamedEventDefinition
 )
 
-from ..specs.BpmnSpecMixin import BpmnSpecMixin
-from ...operators import Attrib, PathAttrib
-
-
-class BpmnDataConverter(DictionaryConverter):
-    """
-    The default converter for task and workflow data.  It allows some commonly used python objects
-    to be converted to a form that can be serialized with JSOM
-
-    It also serves as a simple example for anyone who needs custom data serialization.  If you have
-    custom objects or python objects not included here in your workflow/task data, then you should
-    replace or extend this with one that can handle the contents of your workflow.
-    """
-    def __init__(self):
-
-        super().__init__()
-        self.register(UUID, lambda v: { 'value': str(v) }, lambda v: UUID(v['value']))
-        self.register(datetime, lambda v:  { 'value': v.isoformat() }, lambda v: datetime.fromisoformat(v['value']))
-        self.register(timedelta, lambda v: { 'days': v.days, 'seconds': v.seconds }, lambda v: timedelta(**v))
-
-    def convert(self, obj):
-        self.clean(obj)
-        return super().convert(obj)
-
-    def clean(self, obj):
-        # This removes functions and other callables from task data.
-        # By default we don't want to serialize these
-        if isinstance(obj, dict):
-            items = [ (k, v) for k, v in obj.items() ]
-            for key, value in items:
-                if callable(value):
-                    del obj[key]
-
+from ...specs.BpmnSpecMixin import BpmnSpecMixin
+from ....operators import Attrib, PathAttrib
 
 class BpmnDataSpecificationConverter:
 
@@ -298,52 +264,3 @@ class TaskSpecConverter(DictionaryConverter):
 
     def attrib_from_dict(self, attrib_class, dct):
         return attrib_class(dct['name'])
-
-
-class WorkflowSpecConverter(DictionaryConverter):
-    """
-    This is the base converter for a BPMN workflow spec.
-
-    It will register converters for the task spec types contained in the workflow, as well as
-    the workflow spec class itself.
-
-    This class can be extended if you implement a custom workflow spec type.  See the converter
-    in `workflow_spec_converter` for an example.
-    """
-
-    def __init__(self, spec_class, task_spec_converters, data_converter=None):
-        """
-        Converter for a BPMN workflow spec class.
-
-        The `to_dict` and `from_dict` methods of the given task spec converter classes will
-        be registered, so that they can be restored automatically.
-
-        The data_converter applied to task *spec* data, not task data, and may be `None`.  See
-        `BpmnTaskSpecConverter` for more discussion.
-
-        :param spec_class: the workflow spec class
-        :param task_spec_converters: a list of `BpmnTaskSpecConverter` classes
-        :param data_converter: an optional data converter
-        """
-        super().__init__()
-        self.spec_class = spec_class
-        self.data_converter = data_converter
-
-        self.register(spec_class, self.to_dict, self.from_dict)
-        for converter in task_spec_converters:
-            self.register(converter.spec_class, converter.to_dict, converter.from_dict, converter.typename)
-        self.register(BpmnDataSpecification, BpmnDataSpecificationConverter.to_dict, BpmnDataSpecificationConverter.from_dict)
-
-    def to_dict(self, spec):
-        """
-        The convert method that will be called when a Workflow Spec Converter is registered with a
-        Workflow Converter.
-        """
-        raise NotImplementedError
-
-    def from_dict(self, dct):
-        """
-        The restore method that will be called when a Workflow Spec Converter is registered with a
-        Workflow Converter.
-        """
-        raise NotImplementedError
